backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.5.3
python-dotenv==1.0.0
pillow==10.1.0
bcrypt==4.0.1